refactor code in main.rb and create a class in a file  handle_option
