+++ main.rb
@@ -1,5 +1,6 @@
 require_relative 'app'
 require_relative 'display_menu'
+require_relative 'handle_option'
 
 def main
   books = []
@@ -15,25 +16,6 @@ def main
     handle_option(option, books, people, rentals)
 
     break if option == 7
-  end
-end
-
-def handle_option(option, books, people, rentals)
-  option_actions = {
-    1 => -> { list_all_books(books) },
-    2 => -> { list_all_people(people) },
-    3 => -> { create_person(people) },
-    4 => -> { create_book(books) },
-    5 => -> { create_rental(people, books, rentals) },
-    6 => -> { list_rentals(rentals) },
-    7 => -> { quit }
-  }
-
-  action = option_actions[option]
-  if action
-    action.call
-  else
-    puts 'Invalid option!'
   end
 end
